old-tests/submission/W3C/domhtml/HTMLOptGroupElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OPTGROUP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-optgroup-element.html#dom-optgroup-label" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="test.pl" id="Form1" method="post">
<p>
<select name="select2">
<optgroup label="Regular Employees">
<option>EMP0001</option>
<option>EMP0002</option>
<option>EMP0003A</option>
</optgroup>
<optgroup disabled="disabled" label="Temporary Employees">
<option>EMP0004</option>
<option>EMP0005</option>
</optgroup>
</select>
</p>
</form>



<script>
var domtest = async_test("The label attribute specifies the label assigned to this option group.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vlabel;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("optgroup");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(0);
      vlabel = testNode.label;

      assert_equals(vlabel,"Regular Employees");
       
    domtest.done();
}));

</script>
<p>The label attribute specifies the label assigned to this option group.

    Retrieve the label attribute and examine its value.</p>
<div id="log"></div>
</body></html>
